url/urlsearchparams-getall.html

<!doctype html>
<html>
<head>
<meta charset="utf8">
<link rel="help" href="http://url.spec.whatwg.org/#dom-urlsearchparams-getAll">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
    var params = new URLSearchParams('a=b&c=d');
    assert_array_equals(params.getAll('a'), ['b']);
    assert_array_equals(params.getAll('c'), ['d']);
    assert_array_equals(params.getAll('e'), []);
    params = new URLSearchParams('a=b&c=d&a=e');
    assert_array_equals(params.getAll('a'), ['b', 'e']);
    params = new URLSearchParams('=b&c=d');
    assert_array_equals(params.getAll(''), ['b']);
    params = new URLSearchParams('a=&c=d&a=e');
    assert_array_equals(params.getAll('a'), ['', 'e']);
}, 'getAll() basics');

test(function() {
    var params = new URLSearchParams('a=1&a=2&a=3&a');
    assert_true(params.has('a'), 'Search params object doesn\'t have name "a"');
    var matches = params.getAll('a');
    assert_equals(matches.length, 4, `Unexpected length of name "a" values in
        search params object: ${matches.length}`);
    assert_array_equals(matches, ['1', '2', '3', ''], `Unexpected name "a"
        values: ${matches}`);
    params.set('a', 'one');
    assert_equals(params.get('a'), 'one', 'Search params object doesn\'t ' +
        'have name "a" with value "one"');
    matches = params.getAll('a');
    assert_equals(matches.length, 1, `Unexpected length of name "a" values in
        search params object: ${matches.length}`);
    assert_array_equals(matches, ['one'], `Unexpected name "a" values:
        ${matches}`);
}, 'getAll() multiples');
</script>
</head>
</html>
